requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn>=20.1.0
django-cors-headers>=4.7.0

